feat: change filter logics, refactor code
